src/taia.rs: Consolidate from Corroded `taia_*.rs`
Consolidates all Corrode-translated source code for TAI64NA-related
functionality into a common `taia` module.

Moves all `taia_*()` into an `impl taia` block, as these functions
should eventually all accept `self`.

Code is otherwise unchanged from the original Corrode output.
